pallet_base_fee::pallet

Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event> + IsType<<Self as Config>::RuntimeEvent>;
    type Threshold: BaseFeeThreshold;
    type DefaultBaseFeePerGas: Get<U256>;
    type DefaultElasticity: Get<Permill>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event> + IsType<<Self as Config>::RuntimeEvent>

type Threshold: BaseFeeThreshold

Lower and upper bounds for increasing / decreasing BaseFeePerGas.

type DefaultBaseFeePerGas: Get<U256>

type DefaultElasticity: Get<Permill>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.